Code example for grid scroll listener
